+++ src/main/java/com/iliasdev/dto/ExchangeCurrencyRequestDto.java
@@ -3,10 +3,12 @@ package com.iliasdev.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class ExchangeCurrencyRequestDto {
     private String baseCurrencyCode;
     private String targetCurrencyCode;
-    private double amount;
+    private BigDecimal amount;
 }
