+++ src/main/java/com/iliasdev/model/ExchangeRates.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -11,9 +13,9 @@ public class ExchangeRates {
     private int id;
     private CurrencyModel baseCurrency;
     private CurrencyModel targetCurrency;
-    private double rate;
+    private BigDecimal rate;
 
-    public ExchangeRates(CurrencyModel baseCurrency, CurrencyModel targetCurrency, double rate) {
+    public ExchangeRates(CurrencyModel baseCurrency, CurrencyModel targetCurrency, BigDecimal rate) {
         this.baseCurrency = baseCurrency;
         this.targetCurrency = targetCurrency;
         this.rate = rate;
